% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictInputConfig}
\alias{BatchPredictInputConfig}
\title{BatchPredictInputConfig Object}
\usage{
BatchPredictInputConfig(gcsSource = NULL)
}
\arguments{
\item{gcsSource}{Required}
}
\value{
BatchPredictInputConfig object
}
\description{
BatchPredictInputConfig Object
}
\details{

}
\concept{BatchPredictInputConfig functions}
