% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictOperationMetadata}
\alias{BatchPredictOperationMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
BatchPredictOperationMetadata(outputInfo = NULL, inputConfig = NULL)
}
\arguments{
\item{outputInfo}{Output only}

\item{inputConfig}{Output only}
}
\value{
BatchPredictOperationMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2020-03-16 17:14:06
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BatchPredictOperationMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of BatchPredict operation.
}
\concept{BatchPredictOperationMetadata functions}
