% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictOutputInfo}
\alias{BatchPredictOutputInfo}
\title{BatchPredictOutputInfo Object}
\usage{
BatchPredictOutputInfo(gcsOutputDirectory = NULL)
}
\arguments{
\item{gcsOutputDirectory}{The full path of the Google Cloud Storage directory created, into which the prediction output is written}
}
\value{
BatchPredictOutputInfo object
}
\description{
BatchPredictOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Further describes this batch predict's output. Supplements BatchPredictOutputConfig.
}
\concept{BatchPredictOutputInfo functions}
