% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictResult.metadata}
\alias{BatchPredictResult.metadata}
\title{BatchPredictResult.metadata Object}
\usage{
BatchPredictResult.metadata()
}
\value{
BatchPredictResult.metadata object
}
\description{
BatchPredictResult.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific prediction response metadata.<h4>AutoML Vision Object Detection</h4>`max_bounding_box_count`: (int64) The maximum number of bounding boxes returned per image.<h4>AutoML Video Intelligence Object Tracking</h4>`max_bounding_box_count`: (int64) The maximum number of bounding boxes returned per frame.
}
\seealso{
Other BatchPredictResult functions: \code{\link{BatchPredictResult}}
}
\concept{BatchPredictResult functions}
