% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictResult}
\alias{BatchPredictResult}
\title{BatchPredictResult Object}
\usage{
BatchPredictResult(BatchPredictResult.metadata = NULL, metadata = NULL)
}
\arguments{
\item{BatchPredictResult.metadata}{The \link{BatchPredictResult.metadata} object or list of objects}

\item{metadata}{Additional domain-specific prediction response metadata}
}
\value{
BatchPredictResult object
}
\description{
BatchPredictResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result of the Batch Predict. This message is returned in response of the operation returned by the PredictionService.BatchPredict.
}
\seealso{
Other BatchPredictResult functions: 
\code{\link{BatchPredictResult.metadata}()}
}
\concept{BatchPredictResult functions}
