% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Binding}
\alias{Binding}
\title{Binding Object}
\usage{
Binding(role = NULL, condition = NULL, members = NULL)
}
\arguments{
\item{role}{Role that is assigned to `members`}

\item{condition}{The condition that is associated with this binding}

\item{members}{Specifies the identities requesting access for a Cloud Platform resource}
}
\value{
Binding object
}
\description{
Binding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Associates `members` with a `role`.
}
\concept{Binding functions}
