% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BoundingBoxMetricsEntryConfidenceMetricsEntry}
\alias{BoundingBoxMetricsEntryConfidenceMetricsEntry}
\title{BoundingBoxMetricsEntryConfidenceMetricsEntry Object}
\usage{
BoundingBoxMetricsEntryConfidenceMetricsEntry(
  precision = NULL,
  confidenceThreshold = NULL,
  f1Score = NULL,
  recall = NULL
)
}
\arguments{
\item{precision}{Output only}

\item{confidenceThreshold}{Output only}

\item{f1Score}{Output only}

\item{recall}{Output only}
}
\value{
BoundingBoxMetricsEntryConfidenceMetricsEntry object
}
\description{
BoundingBoxMetricsEntryConfidenceMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for a single confidence threshold.
}
\concept{BoundingBoxMetricsEntryConfidenceMetricsEntry functions}
