% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CreateDatasetOperationMetadata}
\alias{CreateDatasetOperationMetadata}
\title{CreateDatasetOperationMetadata Object}
\usage{
CreateDatasetOperationMetadata()
}
\value{
CreateDatasetOperationMetadata object
}
\description{
CreateDatasetOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of CreateDataset operation.
}
\concept{CreateDatasetOperationMetadata functions}
