% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Dataset.labels}
\alias{Dataset.labels}
\title{Dataset.labels Object}
\usage{
Dataset.labels()
}
\value{
Dataset.labels object
}
\description{
Dataset.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. The labels with user-defined metadata to organize your dataset. Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. Label values are optional. Label keys must start with a letter. See https://goo.gl/xmQnxf for more information on and examples of labels.
}
\seealso{
Other Dataset functions: 
\code{\link{Dataset}()},
\code{\link{projects.locations.datasets.create}()},
\code{\link{projects.locations.datasets.patch}()}
}
\concept{Dataset functions}
