% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(Dataset.labels = NULL, createTime = NULL, labels = NULL,
  textExtractionDatasetMetadata = NULL, description = NULL,
  imageClassificationDatasetMetadata = NULL, etag = NULL,
  textSentimentDatasetMetadata = NULL, name = NULL,
  imageObjectDetectionDatasetMetadata = NULL,
  translationDatasetMetadata = NULL, exampleCount = NULL,
  displayName = NULL, textClassificationDatasetMetadata = NULL)
}
\arguments{
\item{Dataset.labels}{The \link{Dataset.labels} object or list of objects}

\item{createTime}{Output only}

\item{labels}{Optional}

\item{textExtractionDatasetMetadata}{Metadata for a dataset used for text extraction}

\item{description}{User-provided description of the dataset}

\item{imageClassificationDatasetMetadata}{Metadata for a dataset used for image classification}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{textSentimentDatasetMetadata}{Metadata for a dataset used for text sentiment}

\item{name}{Output only}

\item{imageObjectDetectionDatasetMetadata}{Metadata for a dataset used for image object detection}

\item{translationDatasetMetadata}{Metadata for a dataset used for translation}

\item{exampleCount}{Output only}

\item{displayName}{Required}

\item{textClassificationDatasetMetadata}{Metadata for a dataset used for text classification}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A workspace for solving a single, particular machine learning (ML) problem.A workspace contains examples that may be annotated.
}
\seealso{
Other Dataset functions: \code{\link{Dataset.labels}},
  \code{\link{projects.locations.datasets.create}},
  \code{\link{projects.locations.datasets.patch}}
}
\concept{Dataset functions}
