% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExamplePayload}
\alias{ExamplePayload}
\title{ExamplePayload Object}
\usage{
ExamplePayload(document = NULL, image = NULL, textSnippet = NULL)
}
\arguments{
\item{document}{Example document}

\item{image}{Example image}

\item{textSnippet}{Example text}
}
\value{
ExamplePayload object
}
\description{
ExamplePayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Example data used for training or prediction.
}
\concept{ExamplePayload functions}
