% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationDatasetMetadata}
\alias{ImageClassificationDatasetMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
ImageClassificationDatasetMetadata(classificationType = NULL)
}
\arguments{
\item{classificationType}{Required}
}
\value{
ImageClassificationDatasetMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2020-03-30 21:01:47
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ImageClassificationDatasetMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata that is specific to image classification.
}
\concept{ImageClassificationDatasetMetadata functions}
