% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelDeploymentMetadata}
\alias{ImageClassificationModelDeploymentMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
ImageClassificationModelDeploymentMetadata(nodeCount = NULL)
}
\arguments{
\item{nodeCount}{Input only}
}
\value{
ImageClassificationModelDeploymentMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2021-09-09 11:16:19
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ImageClassificationModelDeploymentMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model deployment metadata specific to Image Classification.
}
\concept{ImageClassificationModelDeploymentMetadata functions}
