% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelMetadata}
\alias{ImageClassificationModelMetadata}
\title{ImageClassificationModelMetadata Object}
\usage{
ImageClassificationModelMetadata(nodeQps = NULL,
  trainCostMilliNodeHours = NULL, nodeCount = NULL,
  baseModelId = NULL, modelType = NULL,
  trainBudgetMilliNodeHours = NULL, stopReason = NULL)
}
\arguments{
\item{nodeQps}{Output only}

\item{trainCostMilliNodeHours}{Output only}

\item{nodeCount}{Output only}

\item{baseModelId}{Optional}

\item{modelType}{Optional}

\item{trainBudgetMilliNodeHours}{Optional}

\item{stopReason}{Output only}
}
\value{
ImageClassificationModelMetadata object
}
\description{
ImageClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata for image classification.
}
\concept{ImageClassificationModelMetadata functions}
