% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionEvaluationMetrics}
\alias{ImageObjectDetectionEvaluationMetrics}
\title{ImageObjectDetectionEvaluationMetrics Object}
\usage{
ImageObjectDetectionEvaluationMetrics(
  evaluatedBoundingBoxCount = NULL,
  boundingBoxMeanAveragePrecision = NULL,
  boundingBoxMetricsEntries = NULL
)
}
\arguments{
\item{evaluatedBoundingBoxCount}{Output only}

\item{boundingBoxMeanAveragePrecision}{Output only}

\item{boundingBoxMetricsEntries}{Output only}
}
\value{
ImageObjectDetectionEvaluationMetrics object
}
\description{
ImageObjectDetectionEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for image object detection problems. Evaluates prediction quality of labeled bounding boxes.
}
\concept{ImageObjectDetectionEvaluationMetrics functions}
