% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionModelMetadata}
\alias{ImageObjectDetectionModelMetadata}
\title{ImageObjectDetectionModelMetadata Object}
\usage{
ImageObjectDetectionModelMetadata(
  modelType = NULL,
  trainBudgetMilliNodeHours = NULL
)
}
\arguments{
\item{modelType}{Optional}

\item{trainBudgetMilliNodeHours}{Optional}
}
\value{
ImageObjectDetectionModelMetadata object
}
\description{
ImageObjectDetectionModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to image object detection.
}
\concept{ImageObjectDetectionModelMetadata functions}
