% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{InputConfig.params}
\alias{InputConfig.params}
\title{InputConfig.params Object}
\usage{
InputConfig.params()
}
\value{
InputConfig.params object
}
\description{
InputConfig.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific parameters describing the semantic of the imported data, any string must be up to 25000 characters long. AutoML Tables `schema_inference_version` : (integer) This value must be supplied. The version of the algorithm to use for the initial inference of the column data types of the imported table. Allowed values: '1'.
}
\seealso{
Other InputConfig functions: 
\code{\link{InputConfig}()}
}
\concept{InputConfig functions}
