% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListModelsResponse}
\alias{ListModelsResponse}
\title{ListModelsResponse Object}
\usage{
ListModelsResponse(model = NULL, nextPageToken = NULL)
}
\arguments{
\item{model}{List of models in the requested page}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListModelsResponse object
}
\description{
ListModelsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListModels.
}
\concept{ListModelsResponse functions}
