% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(
  Operation.response = NULL,
  Operation.metadata = NULL,
  name = NULL,
  done = NULL,
  error = NULL,
  response = NULL,
  metadata = NULL
)
}
\arguments{
\item{Operation.response}{The \link{Operation.response} object or list of objects}

\item{Operation.metadata}{The \link{Operation.metadata} object or list of objects}

\item{name}{The server-assigned name, which is only unique within the same service that originally returns it}

\item{done}{If the value is `false`, it means the operation is still in progress}

\item{error}{The error result of the operation in case of failure or cancellation}

\item{response}{The normal response of the operation in case of success}

\item{metadata}{Service-specific metadata associated with the operation}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents a long-running operation that is the result of a network API call.
}
\seealso{
Other Operation functions: 
\code{\link{Operation.metadata}()},
\code{\link{Operation.response}()}
}
\concept{Operation functions}
