% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictRequest.params}
\alias{PredictRequest.params}
\title{PredictRequest.params Object}
\usage{
PredictRequest.params()
}
\value{
PredictRequest.params object
}
\description{
PredictRequest.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific parameters, any string must be up to 25000 characters long. AutoML Vision Classification `score_threshold` : (float) A value from 0.0 to 1.0. When the model makes predictions for an image, it will only produce results that have at least this confidence score. The default is 0.5. AutoML Vision Object Detection `score_threshold` : (float) When Model detects objects on the image, it will only produce bounding boxes which have at least this confidence score. Value in 0 to 1 range, default is 0.5. `max_bounding_box_count` : (int64) The maximum number of bounding boxes returned. The default is 100. The number of returned bounding boxes might be limited by the server. AutoML Tables `feature_importance` : (boolean) Whether feature_importance is populated in the returned list of TablesAnnotation objects. The default is false.
}
\seealso{
Other PredictRequest functions: 
\code{\link{PredictRequest}()},
\code{\link{projects.locations.models.predict}()}
}
\concept{PredictRequest functions}
