% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictResponse.metadata}
\alias{PredictResponse.metadata}
\title{PredictResponse.metadata Object}
\usage{
PredictResponse.metadata()
}
\value{
PredictResponse.metadata object
}
\description{
PredictResponse.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific prediction response metadata. AutoML Vision Object Detection `max_bounding_box_count` : (int64) The maximum number of bounding boxes to return per image. AutoML Natural Language Sentiment Analysis `sentiment_score` : (float, deprecated) A value between -1 and 1, -1 maps to least positive sentiment, while 1 maps to the most positive one and the higher the score, the more positive the sentiment in the document is. Yet these values are relative to the training data, so e.g. if all data was positive then -1 is also positive (though the least). `sentiment_score` is not the same as 'score' and 'magnitude' from Sentiment Analysis in the Natural Language API.
}
\seealso{
Other PredictResponse functions: 
\code{\link{PredictResponse}()}
}
\concept{PredictResponse functions}
