% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_functions.R
\name{projects.locations.models.predict}
\alias{projects.locations.models.predict}
\title{Perform an online prediction. The prediction result is directlyreturned in the response.Available for following ML scenarios, and their expected request payloads:<table><tr><td>AutoML Vision Classification</td><td>An image in .JPEG, .GIF or .PNG format, image_bytes up to 30MB.</td></tr><tr><td>AutoML Vision Object Detection</td><td>An image in .JPEG, .GIF or .PNG format, image_bytes up to 30MB.</td></tr><tr><td>AutoML Natural Language Classification</td><td>A TextSnippet up to 60,000 characters, UTF-8 encoded or a document in.PDF, .TIF or .TIFF format with size upto 2MB.</td></tr><tr><td>AutoML Natural Language Entity Extraction</td><td>A TextSnippet up to 10,000 characters, UTF-8 NFC encoded or a document in .PDF, .TIF or .TIFF format with size upto 20MB.</td></tr><tr><td>AutoML Natural Language Sentiment Analysis</td><td>A TextSnippet up to 60,000 characters, UTF-8 encoded or a document in.PDF, .TIF or .TIFF format with size upto 2MB.</td></tr><tr><td>AutoML Translation</td><td>A TextSnippet up to 25,000 characters, UTF-8 encoded.</td></tr><tr><td>AutoML Tables</td><td>A row with column values matching  the columns of the model, up to 5MB. Not available for FORECASTING  `prediction_type`.</td></tr></table>}
\usage{
projects.locations.models.predict(PredictRequest, name)
}
\arguments{
\item{PredictRequest}{The \link{PredictRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/automl}{Google Documentation}

Other PredictRequest functions: \code{\link{PredictRequest.params}},
  \code{\link{PredictRequest}}
}
\concept{PredictRequest functions}
