% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionAnnotation}
\alias{TextExtractionAnnotation}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
TextExtractionAnnotation(score = NULL, textSegment = NULL)
}
\arguments{
\item{score}{Output only}

\item{textSegment}{An entity annotation will set this, which is the part of the original text to which the annotation pertains}
}
\value{
TextExtractionAnnotation object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2021-04-16 10:03:03
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
TextExtractionAnnotation Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation for identifying spans of text.
}
\concept{TextExtractionAnnotation functions}
