% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TranslationDatasetMetadata}
\alias{TranslationDatasetMetadata}
\title{TranslationDatasetMetadata Object}
\usage{
TranslationDatasetMetadata(
  sourceLanguageCode = NULL,
  targetLanguageCode = NULL
)
}
\arguments{
\item{sourceLanguageCode}{Required}

\item{targetLanguageCode}{Required}
}
\value{
TranslationDatasetMetadata object
}
\description{
TranslationDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata that is specific to translation.
}
\concept{TranslationDatasetMetadata functions}
