% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TranslationEvaluationMetrics}
\alias{TranslationEvaluationMetrics}
\title{TranslationEvaluationMetrics Object}
\usage{
TranslationEvaluationMetrics(baseBleuScore = NULL, bleuScore = NULL)
}
\arguments{
\item{baseBleuScore}{Output only}

\item{bleuScore}{Output only}
}
\value{
TranslationEvaluationMetrics object
}
\description{
TranslationEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Evaluation metrics for the dataset.
}
\concept{TranslationEvaluationMetrics functions}
