% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{AnnotationPayload}
\alias{AnnotationPayload}
\title{AnnotationPayload Object}
\usage{
AnnotationPayload(annotationSpecId = NULL, classification = NULL,
  textExtraction = NULL, imageObjectDetection = NULL,
  displayName = NULL, translation = NULL, textSentiment = NULL)
}
\arguments{
\item{annotationSpecId}{Output only}

\item{classification}{Annotation details for content or image classification}

\item{textExtraction}{Annotation details for text extraction}

\item{imageObjectDetection}{Annotation details for image object detection}

\item{displayName}{Output only}

\item{translation}{Annotation details for translation}

\item{textSentiment}{Annotation details for text sentiment}
}
\value{
AnnotationPayload object
}
\description{
AnnotationPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation information that is relevant to AutoML.
}
\concept{AnnotationPayload functions}
