% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Expr}
\alias{Expr}
\title{Expr Object}
\usage{
Expr(title = NULL, location = NULL, description = NULL,
  expression = NULL)
}
\arguments{
\item{title}{Optional}

\item{location}{Optional}

\item{description}{Optional}

\item{expression}{Textual representation of an expression in Common Expression Language}
}
\value{
Expr object
}
\description{
Expr Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a textual expression in the Common Expression Language (CEL)syntax. CEL is a C-like expression language. The syntax and semantics of CELare documented at https://github.com/google/cel-spec.Example (Comparison):    title: 'Summary size limit'    description: 'Determines if a summary is less than 100 chars'    expression: 'document.summary.size() < 100'Example (Equality):    title: 'Requestor is owner'    description: 'Determines if requestor is the document owner'    expression: 'document.owner == request.auth.claims.email'Example (Logic):    title: 'Public documents'    description: 'Determine whether the document should be publicly visible'    expression: 'document.type != 'private' && document.type != 'internal''Example (Data Manipulation):    title: 'Notification string'    description: 'Create a notification string with a timestamp.'    expression: ''New message received at ' + string(document.create_time)'The exact variables and functions that may be referenced within an expressionare determined by the service that evaluates it. See the servicedocumentation for additional information.
}
\concept{Expr functions}
