% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelDeploymentMetadata}
\alias{ImageClassificationModelDeploymentMetadata}
\title{ImageClassificationModelDeploymentMetadata Object}
\usage{
ImageClassificationModelDeploymentMetadata(nodeCount = NULL)
}
\arguments{
\item{nodeCount}{Input only}
}
\value{
ImageClassificationModelDeploymentMetadata object
}
\description{
ImageClassificationModelDeploymentMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model deployment metadata specific to Image Classification.
}
\concept{ImageClassificationModelDeploymentMetadata functions}
