% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelMetadata}
\alias{ImageClassificationModelMetadata}
\title{ImageClassificationModelMetadata Object}
\usage{
ImageClassificationModelMetadata(modelType = NULL,
  trainBudgetMilliNodeHours = NULL, stopReason = NULL,
  nodeQps = NULL, trainCostMilliNodeHours = NULL, baseModelId = NULL,
  nodeCount = NULL)
}
\arguments{
\item{modelType}{Optional}

\item{trainBudgetMilliNodeHours}{The train budget of creating this model, expressed in milli node}

\item{stopReason}{Output only}

\item{nodeQps}{Output only}

\item{trainCostMilliNodeHours}{Output only}

\item{baseModelId}{Optional}

\item{nodeCount}{Output only}
}
\value{
ImageClassificationModelMetadata object
}
\description{
ImageClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata for image classification.
}
\concept{ImageClassificationModelMetadata functions}
