% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelMetadata}
\alias{ImageClassificationModelMetadata}
\title{ImageClassificationModelMetadata Object}
\usage{
ImageClassificationModelMetadata(
  trainBudgetMilliNodeHours = NULL,
  modelType = NULL,
  baseModelId = NULL
)
}
\arguments{
\item{trainBudgetMilliNodeHours}{Optional}

\item{modelType}{Optional}

\item{baseModelId}{Optional}
}
\value{
ImageClassificationModelMetadata object
}
\description{
ImageClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata for image classification.
}
\concept{ImageClassificationModelMetadata functions}
