% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Layout}
\alias{Layout}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
Layout(boundingPoly = NULL, textSegmentType = NULL,
  textSegment = NULL, pageNumber = NULL)
}
\arguments{
\item{boundingPoly}{The position of the text_segment in the page}

\item{textSegmentType}{The type of the text_segment in document}

\item{textSegment}{Text Segment that represents a segment in}

\item{pageNumber}{Page number of the text_segment in the original document, starts}
}
\value{
Layout object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-12-10 17:15:25
filename: /Users/justinmarciszewski/Dropbox/dev/R/autoGoogleAPI/googleautomlv1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Layout Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the layout information of a text_segment in the document.
}
\concept{Layout functions}
