% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Layout}
\alias{Layout}
\title{Layout Object}
\usage{
Layout(
  textSegmentType = NULL,
  pageNumber = NULL,
  textSegment = NULL,
  boundingPoly = NULL
)
}
\arguments{
\item{textSegmentType}{The type of the text_segment in document}

\item{pageNumber}{Page number of the text_segment in the original document, starts from 1}

\item{textSegment}{Text Segment that represents a segment in document_text}

\item{boundingPoly}{The position of the text_segment in the page}
}
\value{
Layout object
}
\description{
Layout Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the layout information of a text_segment in the document.
}
\concept{Layout functions}
