% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Model}
\alias{Model}
\title{Model Object}
\usage{
Model(Model.labels = NULL, name = NULL,
  translationModelMetadata = NULL, displayName = NULL,
  deploymentState = NULL, createTime = NULL,
  imageClassificationModelMetadata = NULL, labels = NULL,
  textExtractionModelMetadata = NULL, updateTime = NULL,
  textClassificationModelMetadata = NULL, etag = NULL,
  imageObjectDetectionModelMetadata = NULL,
  textSentimentModelMetadata = NULL, datasetId = NULL)
}
\arguments{
\item{Model.labels}{The \link{Model.labels} object or list of objects}

\item{name}{Output only}

\item{translationModelMetadata}{Metadata for translation models}

\item{displayName}{Required}

\item{deploymentState}{Output only}

\item{createTime}{Output only}

\item{imageClassificationModelMetadata}{Metadata for image classification models}

\item{labels}{Optional}

\item{textExtractionModelMetadata}{Metadata for text extraction models}

\item{updateTime}{Output only}

\item{textClassificationModelMetadata}{Metadata for text classification models}

\item{etag}{Used to perform a consistent read-modify-write updates}

\item{imageObjectDetectionModelMetadata}{Metadata for image object detection models}

\item{textSentimentModelMetadata}{Metadata for text sentiment models}

\item{datasetId}{Required}
}
\value{
Model object
}
\description{
Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
API proto representing a trained machine learning model.
}
\seealso{
Other Model functions: \code{\link{Model.labels}},
  \code{\link{projects.locations.models.create}},
  \code{\link{projects.locations.models.patch}}
}
\concept{Model functions}
