% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Model}
\alias{Model}
\title{Model Object}
\usage{
Model(
  Model.labels = NULL,
  updateTime = NULL,
  displayName = NULL,
  imageObjectDetectionModelMetadata = NULL,
  etag = NULL,
  textSentimentModelMetadata = NULL,
  createTime = NULL,
  deploymentState = NULL,
  imageClassificationModelMetadata = NULL,
  translationModelMetadata = NULL,
  textExtractionModelMetadata = NULL,
  name = NULL,
  labels = NULL,
  datasetId = NULL,
  textClassificationModelMetadata = NULL
)
}
\arguments{
\item{Model.labels}{The \link{Model.labels} object or list of objects}

\item{updateTime}{Output only}

\item{displayName}{Required}

\item{imageObjectDetectionModelMetadata}{Metadata for image object detection models}

\item{etag}{Used to perform a consistent read-modify-write updates}

\item{textSentimentModelMetadata}{Metadata for text sentiment models}

\item{createTime}{Output only}

\item{deploymentState}{Output only}

\item{imageClassificationModelMetadata}{Metadata for image classification models}

\item{translationModelMetadata}{Metadata for translation models}

\item{textExtractionModelMetadata}{Metadata for text extraction models}

\item{name}{Output only}

\item{labels}{Optional}

\item{datasetId}{Required}

\item{textClassificationModelMetadata}{Metadata for text classification models}
}
\value{
Model object
}
\description{
Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
API proto representing a trained machine learning model.
}
\seealso{
Other Model functions: 
\code{\link{Model.labels}()},
\code{\link{projects.locations.models.create}()},
\code{\link{projects.locations.models.patch}()}
}
\concept{Model functions}
