% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ModelEvaluation}
\alias{ModelEvaluation}
\title{ModelEvaluation Object}
\usage{
ModelEvaluation(evaluatedExampleCount = NULL, name = NULL,
  textSentimentEvaluationMetrics = NULL,
  imageObjectDetectionEvaluationMetrics = NULL,
  translationEvaluationMetrics = NULL, annotationSpecId = NULL,
  textExtractionEvaluationMetrics = NULL, displayName = NULL,
  classificationEvaluationMetrics = NULL, createTime = NULL)
}
\arguments{
\item{evaluatedExampleCount}{Output only}

\item{name}{Output only}

\item{textSentimentEvaluationMetrics}{Evaluation metrics for text sentiment models}

\item{imageObjectDetectionEvaluationMetrics}{Model evaluation metrics for image object detection}

\item{translationEvaluationMetrics}{Model evaluation metrics for translation}

\item{annotationSpecId}{Output only}

\item{textExtractionEvaluationMetrics}{Evaluation metrics for text extraction models}

\item{displayName}{Output only}

\item{classificationEvaluationMetrics}{Model evaluation metrics for image, text, video and tables}

\item{createTime}{Output only}
}
\value{
ModelEvaluation object
}
\description{
ModelEvaluation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Evaluation results of a model.
}
\concept{ModelEvaluation functions}
