% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Policy}
\alias{Policy}
\title{Policy Object}
\usage{
Policy(bindings = NULL, etag = NULL, version = NULL)
}
\arguments{
\item{bindings}{Associates a list of `members` to a `role`}

\item{etag}{`etag` is used for optimistic concurrency control as a way to help}

\item{version}{Specifies the format of the policy}
}
\value{
Policy object
}
\description{
Policy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Identity and Access Management (IAM) policy, which specifies accesscontrols for Google Cloud resources.A `Policy` is a collection of `bindings`. A `binding` binds one or more`members` to a single `role`. Members can be user accounts, service accounts,Google groups, and domains (such as G Suite). A `role` is a named list ofpermissions; each `role` can be an IAM predefined role or a user-createdcustom role.Optionally, a `binding` can specify a `condition`, which is a logicalexpression that allows access to a resource only if the expression evaluatesto `true`. A condition can add constraints based on attributes of therequest, the resource, or both.**JSON example:**    {      'bindings': [        {          'role': 'roles/resourcemanager.organizationAdmin',          'members': [            'user:mike@example.com',            'group:admins@example.com',            'domain:google.com',            'serviceAccount:my-project-id@appspot.gserviceaccount.com'          ]        },        {          'role': 'roles/resourcemanager.organizationViewer',          'members': ['user:eve@example.com'],          'condition': {            'title': 'expirable access',            'description': 'Does not grant access after Sep 2020',            'expression': 'request.time < timestamp('2020-10-01T00:00:00.000Z')',          }        }      ],      'etag': 'BwWWja0YfJA=',      'version': 3    }**YAML example:**    bindings:    - members:      - user:mike@example.com      - group:admins@example.com      - domain:google.com      - serviceAccount:my-project-id@appspot.gserviceaccount.com      role: roles/resourcemanager.organizationAdmin    - members:      - user:eve@example.com      role: roles/resourcemanager.organizationViewer      condition:        title: expirable access        description: Does not grant access after Sep 2020        expression: request.time < timestamp('2020-10-01T00:00:00.000Z')    - etag: BwWWja0YfJA=    - version: 3For a description of IAM and its features, see the[IAM documentation](https://cloud.google.com/iam/docs/).
}
\concept{Policy functions}
