% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionAnnotation}
\alias{TextExtractionAnnotation}
\title{TextExtractionAnnotation Object}
\usage{
TextExtractionAnnotation(textSegment = NULL, score = NULL)
}
\arguments{
\item{textSegment}{An entity annotation will set this, which is the part of the original text to which the annotation pertains}

\item{score}{Output only}
}
\value{
TextExtractionAnnotation object
}
\description{
TextExtractionAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation for identifying spans of text.
}
\concept{TextExtractionAnnotation functions}
