% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSnippet}
\alias{TextSnippet}
\title{TextSnippet Object}
\usage{
TextSnippet(mimeType = NULL, contentUri = NULL, content = NULL)
}
\arguments{
\item{mimeType}{Optional}

\item{contentUri}{Output only}

\item{content}{Required}
}
\value{
TextSnippet object
}
\description{
TextSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a text snippet.
}
\concept{TextSnippet functions}
